Added a Ruby implementation of the quicksort algo
This version alters the array in-place, without any temporary arrays, so it has a space complexity of O(log(n)), due basically to the recursion stack. Also, with the pivot selected being always the middle element, it does not degrade to O(n^2) when the array is already sorted. This way, the time complexity should be on the average case of O(n * log(n))